requirements.txt
Django
django-cors-headers
django-extensions
djangorestframework
PyJWT
six
psycopg2
drf-yasg
setuptools